+++ oo/src/language/analyzing/stree.rs
@@ -9,7 +9,7 @@ pub enum STree {
     START { functions: Vec<STree> },
     FUNCTION { function_type: FunctionType, return_type: VariableType, name: String, params: Vec<(String, VariableType)>, body: Box<STree> },
     BLOCK { statements: Vec<STree> },
-    LET_STMT { id: String, var_type: VariableType, expression: Option<Box<STree>> },
+    LET_STMT { id: String, var_type: VariableType, mutable: bool, expression: Option<Box<STree>> },
     ASSIGN_STMT { id: String, expression: Box<STree> },
     FOR_EXPR { init: Option<Box<STree>>, condition: Box<STree>, modifier: Option<Box<STree>>, body: Box<STree> },
     FOR_EACH { variable: String, iterable: Box<STree>, body: Box<STree> },